refactor: use sglang.utils.is_in_ci from jit_kernel and benchmarks
Remove duplicate is_in_ci implementations from jit_kernel.utils and
jit_kernel.benchmark.utils; route get_*_range helpers and benchmark
scripts through envs.SGLANG_IS_IN_CI (sglang.utils.is_in_ci).

Made-with: Cursor
