+++ python/sglang/jit_kernel/benchmark/bench_fused_norm_scale_shift.py
@@ -8,13 +8,14 @@ import torch
 import triton
 import triton.testing
 
-from sglang.jit_kernel.benchmark.utils import is_in_ci, run_benchmark_no_cudagraph
+from sglang.jit_kernel.benchmark.utils import run_benchmark_no_cudagraph
 from sglang.multimodal_gen.runtime.layers.layernorm import (
     LayerNormScaleShift,
     RMSNormScaleShift,
     ScaleResidualLayerNormScaleShift,
     ScaleResidualRMSNormScaleShift,
 )
+from sglang.utils import is_in_ci
 
 if is_in_ci():
     B_RANGE, S_RANGE, D_RANGE = [1], [128], [1024]
