+++ python/sglang/jit_kernel/benchmark/bench_per_token_group_quant_8bit.py
@@ -5,10 +5,7 @@ import torch
 import triton
 from sgl_kernel.test_utils import create_per_token_group_quant_test_data
 
-from sglang.jit_kernel.benchmark.utils import (
-    get_benchmark_range,
-    is_in_ci,
-)
+from sglang.jit_kernel.benchmark.utils import get_benchmark_range
 from sglang.jit_kernel.per_token_group_quant_8bit import (
     per_token_group_quant_8bit as sglang_per_token_group_quant_8bit,
 )
@@ -20,6 +17,7 @@ from sglang.srt.layers.quantization.fp8_kernel import (
 )
 from sglang.srt.utils import is_hip
 from sglang.srt.utils.bench_utils import bench_kineto
+from sglang.utils import is_in_ci
 
 IS_CI = is_in_ci()
