+++ python/sglang/jit_kernel/benchmark/utils.py
@@ -5,17 +5,12 @@ from typing import Callable, List, Tuple
 import torch
 import triton.testing
 
-from sglang.jit_kernel.utils import is_in_ci as jit_kernel_is_in_ci
+from sglang.utils import is_in_ci
 
 # Common constants
 DEFAULT_DTYPE = torch.bfloat16
 DEFAULT_DEVICE = "cuda"
 DEFAULT_QUANTILES = [0.5, 0.2, 0.8]
-
-
-def is_in_ci() -> bool:
-    """Check if running in CI environment."""
-    return jit_kernel_is_in_ci()
 
 
 def get_benchmark_range(full_range: List, ci_range: List) -> List:
