+++ python/sglang/jit_kernel/utils.py
@@ -7,16 +7,13 @@ from typing import TYPE_CHECKING, Any, Callable, List, Tuple, TypeAlias, TypeVar
 
 import torch
 
+from sglang.utils import is_in_ci
+
 if TYPE_CHECKING:
     from tvm_ffi import Module
 
 F = TypeVar("F", bound=Callable[..., Any])
 _FULL_TEST_ENV_VAR = "SGLANG_JIT_KERNEL_RUN_FULL_TESTS"
-
-
-def is_in_ci() -> bool:
-    ci_env_vars = ("SGLANG_IS_IN_CI", "CI", "GITHUB_ACTIONS")
-    return any(os.getenv(env_var, "false").lower() == "true" for env_var in ci_env_vars)
 
 
 def should_run_full_tests() -> bool:
